- skydomes from hdri - now the framebuffers are in 32 bits floating point to avoid bugs with hdri values - back face culling

+++ Source/Framebuffer.cpp
@@ -16,7 +16,7 @@ void Framebuffer::Init(float width, float height)
     glGenTextures(1, &m_ColorBufferTexture);
     glBindTexture(GL_TEXTURE_2D, m_ColorBufferTexture);
 
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, width, height, 0, GL_RGBA, GL_FLOAT, nullptr);
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, width, height, 0, GL_RGBA, GL_FLOAT, nullptr);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
